Moved AlbumReleaseDateEnhancerAfterConvertCallback to 'read' package
