+++ src/test/java/com/odeyalo/sonata/harmony/repository/R2dbcAlbumReleaseRepositoryTest.java
@@ -4,7 +4,7 @@ import com.odeyalo.sonata.harmony.entity.AlbumReleaseEntity;
 import com.odeyalo.sonata.harmony.model.AlbumType;
 import com.odeyalo.sonata.harmony.model.ReleaseDate;
 import com.odeyalo.sonata.harmony.repository.r2dbc.callback.write.ReleaseDateAssociationReleaseAlbumEntityBeforeSaveCallback;
-import com.odeyalo.sonata.harmony.repository.r2dbc.callback.write.AlbumReleaseDateEnhancerAfterConvertCallback;
+import com.odeyalo.sonata.harmony.repository.r2dbc.callback.read.AlbumReleaseDateEnhancerAfterConvertCallback;
 import com.odeyalo.sonata.harmony.repository.r2dbc.delegate.R2dbcAlbumReleaseRepositoryDelegate;
 import com.odeyalo.sonata.harmony.repository.r2dbc.support.release.FormattedString2ReleaseDateConverter;
 import com.odeyalo.sonata.harmony.repository.r2dbc.support.release.ReleaseDateDecoder;
